+++ CHANGELOG.md
@@ -8,6 +8,36 @@ This project adheres to [Semantic Versioning](https://semver.org).
 
 [Commits](https://github.com/thebigmunch/google-music-proto/compare/1.3.0...master)
 
+### Added
+
+* Expose undocumented batch call base classes:
+	* ``ActivityRecord``
+	* ``PlaylistBatch``
+	* ``RadioEditStation``
+	* ``TrackBatch``
+
+### Changed
+
+* Batch calls now use staticmethods to build the events/mutations.
+  See the docs for each call's methods.
+  The events/mutations are passed to the batch call class.
+  This allows for different types of operations to be done in one call.
+
+### Removed
+
+* Batch call subclasses:
+	* ``ActivityRecordPlay``
+	* ``ActivityRecordRate``
+	* ``PlaylistBatchCreate``
+	* ``PlaylistBatchDelete``
+	* ``PlaylistBatchUpdate``
+	* ``RadioEditStationCreateOrGet``
+	* ``RadioEditStationDelete``
+	* ``TrackBatchCreate``
+	* ``TrackBatchDelete``
+	* ``TrackBatchUpdate``
+
+
 
 
 ## [1.3.0](https://github.com/thebigmunch/google-music-proto/releases/tag/1.2.0) (2018-10-25)