+++ CHANGELOG.md
@@ -51,7 +51,7 @@ This project adheres to [Semantic Versioning](https://semver.org).
 ### Added
 
 * Support for I'm Feeling Lucky Radio by ``station_id`` set to ``'IFL'``.
-* ``PlaylistsDelete``.
+* ``PlaylistsDelete``
 
 
 ## [1.2.0](https://github.com/thebigmunch/google-music-proto/releases/tag/1.2.0) (2018-10-23)
@@ -60,7 +60,7 @@ This project adheres to [Semantic Versioning](https://semver.org).
 
 ### Added
 
-* ``ListenNowItemType``.
+* ``ListenNowItemType``
 
 ### Changed