+++ CHANGELOG.md
@@ -8,6 +8,10 @@ This project adheres to [Semantic Versioning](https://semver.org).
 
 [Commits](https://github.com/thebigmunch/google-music-proto/compare/2.0.0...master)
 
+### Added
+
+* ``PlaylistEntriesBatch``
+
 
 
 ## [2.0.0](https://github.com/thebigmunch/google-music-proto/releases/tag/2.0.0) (2018-11-05)