+++ CHANGELOG.md
@@ -12,6 +12,10 @@ This project adheres to [Semantic Versioning](https://semver.org).
 
 * ``PlaylistEntriesBatch``
 
+### Fixed
+
+* ``PlaylistEntriesShared``
+
 
 
 ## [2.0.0](https://github.com/thebigmunch/google-music-proto/releases/tag/2.0.0) (2018-11-05)